kernel-5.10.0-0.rc6.92
Update configs: CONFIG_KCSAN added for x86 configs and CONFIG_KASAN_SW_TAGS
added for aarch64 configs. This is required because between last update and now,
gcc 11 was added to rhel9, and probably because it these new configs are now
visible. As there was no setting, the process_configs.sh script started to
report a failure, resulting in a brew build failure.

Signed-off-by: Herton R. Krzesinski <herton@redhat.com>